Improve make checks, fix annoyances
Without this patch there are a number of annoyances with the make
checks.  This patch fixes the following:

 1. Newlines not being added to auto-generated documentation.
 2. Auto generated documentation failing if README.md does not exist.
 3. Checks execute against .git/ index
 4. terraform validate executes redundantly

To address the doc generation annoyances, this patch converts
combine_docfiles to fairly simple bash.  The behavior is pure sed, the
rest of the bash is argument and error handling.

For the others, the patch refactors make.sh to use a general find |
xargs pattern with a single function to manage exclusion lists.

check_shebangs is removed because `#! /bin/bash -e` should not be used
because it does not take effect if the script is called via bash
<filename>.  Instead, calling `set -e` is preferred.