Manage GCP fixture resources like CI does
Without this patch there isn't a clear way to run integration tests
locally on a workstation reflecting how the CI system runs integration
tests.  This is a problem because it creates drift and divergence from
how CI operates, creating friction in the development process.

This patch addresses the problem by allowing GCP resources to be created
and managed just like they are with the CI system.  A project and
service account are created and provided as inputs to the integration
test suite.