Add make integration_test_run task and basic kitchen test
This patch adds a make integration_test_run task reflecting the CI
integration-test job.  Both the make target and the CI job execute
test/ci_integration.sh inside of a Docker container.  The Make task and
CI Job both share the responsibility of fetching the container, setting
up the context for the tests (Service Account credentials, GCP project
to run in, etc...), and then executing the container and handing off to
ci_integration.sh.

The simple_example kitchen test verifies the default behavior of the
startup script. It loads a basic startup-script-custom script and logs a
message.